moved data fetching to hook
